Push: Merge SharedConfig into Push class
Push is the only user of the SharedConfig class now that
Feedback is gone, so there is no need to keep it separate
anymore.
