+++ ApnsPHP/Tests/PushSendTest.php
@@ -9,6 +9,8 @@
 
 namespace ApnsPHP\Tests;
 
+use stdClass;
+
 /**
  * This class contains tests for the send function
  *
@@ -36,7 +38,7 @@ class PushSendTest extends PushTest
      */
     public function testSendThrowsExceptionOnEmptyQueue()
     {
-        $this->set_reflection_property_value('hSocket', true);
+        $this->set_reflection_property_value('hSocket', new stdClass());
 
         $this->expectException('ApnsPHP\Push\Exception');
         $this->expectExceptionMessage('No notifications queued to be sent');
@@ -51,32 +53,47 @@ class PushSendTest extends PushTest
      */
     public function testSendFailsWithoutRetrying()
     {
-        // because we cannot mock private functions we need to mock a function that httpSend uses
-        // to get it to return the result we need
         $this->mock_function('curl_exec', function () {
             return false;
         });
+        $this->mock_function('curl_setopt_array', function () {
+            return true;
+        });
+        $this->mock_function('curl_getinfo', function () {
+            return 404;
+        });
+        $this->mock_function('curl_close', function () {
+            return null;
+        });
+        $this->mock_function('curl_init', function () {
+            return new stdClass();
+        });
 
+        $error = [
+            'command' => 8,
+            'statusCode' => 4,
+            'identifier' => 1,
+            'time' => 1620029695,
+            'statusMessage' => 'Missing payload'
+        ];
         $message = [ 1 => [ 'MESSAGE' => $this->message, 'ERRORS' => [] ] ];
 
         $this->set_reflection_property_value('environment', 1);
-        $this->set_reflection_property_value('hSocket', curl_init());
+        $this->set_reflection_property_value('hSocket', new stdClass());
         $this->set_reflection_property_value('messageQueue', $message);
+        $this->set_reflection_property_value('logger', $this->logger);
+        $this->set_reflection_property_value('writeInterval', 0);
 
-        $this->class->setWriteInterval(0);
-
-        $sendMessages = [
-            ['Sending messages queue, run #1: 1 message(s) left in queue.'],
-            ['Sending messages queue, run #2: 1 message(s) left in queue.']
-        ];
-
-        $this->class->expects($this->exactly(4))
-                    ->method('logger')
-                    ->will($this->returnValue($this->logger));
-
-        $this->logger->expects($this->exactly(2))
+        $this->logger->expects($this->exactly(6))
                      ->method('info')
-                     ->withConsecutive(...$sendMessages);
+                     ->withConsecutive(
+                         [ 'Sending messages queue, run #1: 1 message(s) left in queue.' ],
+                         [ 'Disconnected.' ],
+                         [ 'Trying to initialize HTTP/2 backend...' ],
+                         [ 'Initializing HTTP/2 backend with certificate.' ],
+                         [ 'Initialized HTTP/2 backend.' ],
+                         [ 'Sending messages queue, run #2: 1 message(s) left in queue.' ],
+                     );
 
         $this->logger->expects($this->once())
                      ->method('debug')
@@ -85,25 +102,15 @@ class PushSendTest extends PushTest
         $this->logger->expects($this->once())
                      ->method('warning')
                      ->with('Message ID 1 [custom identifier: unset] has an unrecoverable error
-                                 (4), removing from queue without retrying...');
-
-        $this->class->expects($this->exactly(1))
-                    ->method('updateQueue')
-                    ->will($this->returnCallback(function () {
-                        $queue = $this->get_accessible_reflection_property('messageQueue')->getValue($this->class);
-                        $queue[1]['ERRORS'][] = [
-                            'command' => 8,
-                            'statusCode' => 4,
-                            'identifier' => 1,
-                            'time' => 1620029695,
-                            'statusMessage' => 'Missing payload'
-                        ];
-                        return true;
-                    }));
+                                 (404), removing from queue without retrying...');
 
         $this->class->send();
 
         $this->unmock_function('curl_exec');
+        $this->unmock_function('curl_setopt_array');
+        $this->unmock_function('curl_getinfo');
+        $this->unmock_function('curl_close');
+        $this->unmock_function('curl_init');
     }
 
     /**
@@ -113,66 +120,70 @@ class PushSendTest extends PushTest
      */
     public function testSendFailsWithRetrying()
     {
-        // because we cannot mock private functions we need to mock a function that httpSend uses
-        // to get it to return the result we need
         $this->mock_function('curl_exec', function () {
             return false;
+        });
+        $this->mock_function('curl_setopt_array', function () {
+            return true;
+        });
+        $this->mock_function('curl_getinfo', function () {
+            return 429;
+        });
+        $this->mock_function('curl_close', function () {
+            return null;
+        });
+        $this->mock_function('curl_init', function () {
+            return new stdClass();
         });
 
         $message = [ 1 => [ 'MESSAGE' => $this->message, 'ERRORS' => [] ] ];
 
         $this->set_reflection_property_value('environment', 1);
-        $this->set_reflection_property_value('hSocket', curl_init());
+        $this->set_reflection_property_value('hSocket', new stdClass());
         $this->set_reflection_property_value('messageQueue', $message);
+        $this->set_reflection_property_value('logger', $this->logger);
+        $this->set_reflection_property_value('writeInterval', 0);
 
-        $this->class->setWriteInterval(0);
-
-        $queueMessages = [
-            ['Sending messages queue, run #1: 1 message(s) left in queue.'],
-            ['Sending messages queue, run #2: 1 message(s) left in queue.'],
-            ['Sending messages queue, run #3: 1 message(s) left in queue.'],
-            ['Sending messages queue, run #4: 1 message(s) left in queue.']
-        ];
-
-        $sendMessages = [
-            ['Sending message ID 1 [custom identifier: unset] (1/3): 0 bytes.'],
-            ['Sending message ID 1 [custom identifier: unset] (2/3): 0 bytes.'],
-            ['Sending message ID 1 [custom identifier: unset] (3/3): 0 bytes.']
-        ];
-
-        $this->class->expects($this->exactly(8))
-                    ->method('logger')
-                    ->will($this->returnValue($this->logger));
-
-        $this->logger->expects($this->exactly(4))
+        $this->logger->expects($this->exactly(16))
                      ->method('info')
-                     ->withConsecutive(...$queueMessages);
+                     ->withConsecutive(
+                         [ 'Sending messages queue, run #1: 1 message(s) left in queue.' ],
+                         [ 'Disconnected.' ],
+                         [ 'Trying to initialize HTTP/2 backend...' ],
+                         [ 'Initializing HTTP/2 backend with certificate.' ],
+                         [ 'Initialized HTTP/2 backend.' ],
+                         [ 'Sending messages queue, run #2: 1 message(s) left in queue.' ],
+                         [ 'Disconnected.' ],
+                         [ 'Trying to initialize HTTP/2 backend...' ],
+                         [ 'Initializing HTTP/2 backend with certificate.' ],
+                         [ 'Initialized HTTP/2 backend.' ],
+                         [ 'Sending messages queue, run #3: 1 message(s) left in queue.' ],
+                         [ 'Disconnected.' ],
+                         [ 'Trying to initialize HTTP/2 backend...' ],
+                         [ 'Initializing HTTP/2 backend with certificate.' ],
+                         [ 'Initialized HTTP/2 backend.' ],
+                         [ 'Sending messages queue, run #4: 1 message(s) left in queue.' ],
+                     );
 
         $this->logger->expects($this->exactly(3))
                      ->method('debug')
-                     ->withConsecutive(...$sendMessages);
+                     ->withConsecutive(
+                         [ 'Sending message ID 1 [custom identifier: unset] (1/3): 0 bytes.' ],
+                         [ 'Sending message ID 1 [custom identifier: unset] (2/3): 0 bytes.' ],
+                         [ 'Sending message ID 1 [custom identifier: unset] (3/3): 0 bytes.' ],
+                     );
 
         $this->logger->expects($this->once())
                      ->method('warning')
                      ->with('Message ID 1 [custom identifier: unset] has 3 errors, removing from queue...');
 
-        $this->class->expects($this->exactly(3))
-                    ->method('updateQueue')
-                    ->will($this->returnCallback(function () {
-                        $queue = $this->get_accessible_reflection_property('messageQueue')->getValue($this->class);
-                        $queue[1]['ERRORS'][] = [
-                            'command' => 8,
-                            'statusCode' => 999,
-                            'identifier' => 1,
-                            'time' => 1620029695,
-                            'statusMessage' => 'Internal error'
-                        ];
-                        return true;
-                    }));
-
         $this->class->send();
 
         $this->unmock_function('curl_exec');
+        $this->unmock_function('curl_setopt_array');
+        $this->unmock_function('curl_getinfo');
+        $this->unmock_function('curl_close');
+        $this->unmock_function('curl_init');
     }
 
     /**
@@ -182,56 +193,57 @@ class PushSendTest extends PushTest
      */
     public function testSendRemovesWhenNoError()
     {
-        // because we cannot mock private functions we need to mock a function that httpSend uses
-        // to get it to return the result we need
         $this->mock_function('curl_exec', function () {
             return false;
+        });
+        $this->mock_function('curl_setopt_array', function () {
+            return true;
+        });
+        $this->mock_function('curl_getinfo', function () {
+            return 200;
+        });
+        $this->mock_function('curl_close', function () {
+            return null;
+        });
+        $this->mock_function('curl_init', function () {
+            return new stdClass();
         });
 
         $message = [ 1 => [ 'MESSAGE' => $this->message, 'ERRORS' => [] ] ];
 
         $this->set_reflection_property_value('environment', 1);
-        $this->set_reflection_property_value('hSocket', curl_init());
+        $this->set_reflection_property_value('hSocket', new stdClass());
         $this->set_reflection_property_value('messageQueue', $message);
+        $this->set_reflection_property_value('logger', $this->logger);
+        $this->set_reflection_property_value('writeInterval', 0);
 
-        $this->class->setWriteInterval(0);
-
-        $queueMessages = [
-            ['Sending messages queue, run #1: 1 message(s) left in queue.'],
-            ['Sending messages queue, run #2: 1 message(s) left in queue.'],
-            ['Message ID 1 [custom identifier: unset] has no error (0),
-                                 removing from queue...'],
-        ];
-
-        $this->class->expects($this->exactly(4))
-                    ->method('logger')
-                    ->will($this->returnValue($this->logger));
-
-        $this->logger->expects($this->exactly(3))
+        $this->logger->expects($this->exactly(7))
                      ->method('info')
-                     ->withConsecutive(...$queueMessages);
+                     ->withConsecutive(
+                         [ 'Sending messages queue, run #1: 1 message(s) left in queue.' ],
+                         [ 'Disconnected.' ],
+                         [ 'Trying to initialize HTTP/2 backend...' ],
+                         [ 'Initializing HTTP/2 backend with certificate.' ],
+                         [ 'Initialized HTTP/2 backend.' ],
+                         [ 'Sending messages queue, run #2: 1 message(s) left in queue.' ],
+                         [ 'Message ID 1 [custom identifier: unset] has no error (200),
+                                 removing from queue...'],
+                     );
 
         $this->logger->expects($this->once())
                      ->method('debug')
                      ->with('Sending message ID 1 [custom identifier: unset] (1/3): 0 bytes.');
 
-        $this->class->expects($this->exactly(1))
-                    ->method('updateQueue')
-                    ->will($this->returnCallback(function () {
-                        $queue = $this->get_accessible_reflection_property('messageQueue')->getValue($this->class);
-                        $queue[1]['ERRORS'][] = [
-                            'command' => 8,
-                            'statusCode' => 0,
-                            'identifier' => 1,
-                            'time' => 1620029695,
-                            'statusMessage' => 'No error'
-                        ];
-                        return true;
-                    }));
+        $this->logger->expects($this->never())
+                     ->method('warning');
 
         $this->class->send();
 
         $this->unmock_function('curl_exec');
+        $this->unmock_function('curl_setopt_array');
+        $this->unmock_function('curl_getinfo');
+        $this->unmock_function('curl_close');
+        $this->unmock_function('curl_init');
     }
 
     /**
@@ -241,23 +253,29 @@ class PushSendTest extends PushTest
      */
     public function testSendSuccessfullySends()
     {
-        // because we cannot mock private functions we need to mock a function that httpSend uses
-        // to get it to return the result we need
         $this->mock_function('curl_exec', function () {
             return true;
+        });
+        $this->mock_function('curl_setopt_array', function () {
+            return true;
+        });
+        $this->mock_function('curl_getinfo', function () {
+            return 200;
+        });
+        $this->mock_function('curl_close', function () {
+            return null;
+        });
+        $this->mock_function('curl_init', function () {
+            return new stdClass();
         });
 
         $message = [ 1 => [ 'MESSAGE' => $this->message, 'ERRORS' => [] ] ];
 
         $this->set_reflection_property_value('environment', 1);
-        $this->set_reflection_property_value('hSocket', curl_init());
+        $this->set_reflection_property_value('hSocket', new stdClass());
         $this->set_reflection_property_value('messageQueue', $message);
-
-        $this->class->setWriteInterval(0);
-
-        $this->class->expects($this->exactly(2))
-                    ->method('logger')
-                    ->will($this->returnValue($this->logger));
+        $this->set_reflection_property_value('logger', $this->logger);
+        $this->set_reflection_property_value('writeInterval', 0);
 
         $this->logger->expects($this->once())
                      ->method('info')
@@ -267,12 +285,12 @@ class PushSendTest extends PushTest
                      ->method('debug')
                      ->with('Sending message ID 1 [custom identifier: unset] (1/3): 0 bytes.');
 
-        $this->class->expects($this->once())
-                    ->method('updateQueue')
-                    ->will($this->returnValue(false));
-
         $this->class->send();
 
         $this->unmock_function('curl_exec');
+        $this->unmock_function('curl_setopt_array');
+        $this->unmock_function('curl_getinfo');
+        $this->unmock_function('curl_close');
+        $this->unmock_function('curl_init');
     }
 }
