+++ phpunit.xml
@@ -9,16 +9,13 @@
 
     <testsuites>
         <testsuite name="Base">
-            <file>ApnsPHP/Tests/SharedConfigTest.php</file>
-            <file>ApnsPHP/Tests/SharedConfigDisconnectTest.php</file>
-            <file>ApnsPHP/Tests/SharedConfigHttpInitTest.php</file>
-            <file>ApnsPHP/Tests/SharedConfigConnectTest.php</file>
-            <file>ApnsPHP/Tests/MessageTest.php</file>
             <file>ApnsPHP/Tests/MessageAddRecipientTest.php</file>
             <file>ApnsPHP/Tests/MessageSelfForRecipientTest.php</file>
             <file>ApnsPHP/Tests/MessageGetPayloadDictionaryTest.php</file>
             <file>ApnsPHP/Tests/MessageGetPayloadTest.php</file>
-            <file>ApnsPHP/Tests/PushTest.php</file>
+            <file>ApnsPHP/Tests/PushDisconnectTest.php</file>
+            <file>ApnsPHP/Tests/PushHttpInitTest.php</file>
+            <file>ApnsPHP/Tests/PushConnectTest.php</file>
             <file>ApnsPHP/Tests/PushAddTest.php</file>
             <file>ApnsPHP/Tests/PushSendTest.php</file>
             <file>ApnsPHP/Tests/PushHttpSendTest.php</file>
